+++ src/Symfony/Component/HttpKernel/Exception/FlattenException.php
@@ -11,8 +11,6 @@
 
 namespace Symfony\Component\HttpKernel\Exception;
 
-use Symfony\Component\Debug\Exception\FlattenException as DebugFlattenException;
-
 /**
  * FlattenException wraps a PHP Exception to be able to serialize it.
  *
@@ -22,6 +20,4 @@ use Symfony\Component\Debug\Exception\FlattenException as DebugFlattenException;
  *
  * @deprecated Deprecated in 2.3, to be removed in 3.0. Use the same class from the Debug component instead.
  */
-class FlattenException extends DebugFlattenException
-{
-}
+class_exists('Symfony\Component\Debug\Exception\FlattenException');