+++ src/Symfony/Component/HttpKernel/Tests/DataCollector/ConfigDataCollectorTest.php
@@ -69,10 +69,6 @@ class KernelForTest extends Kernel
     {
     }
 
-    public function init()
-    {
-    }
-
     public function getBundles()
     {
         return array();