+++ src/Symfony/Bridge/Twig/NodeVisitor/Scope.php
@@ -22,11 +22,6 @@ class Scope
     private $parent;
 
     /**
-     * @var Scope[]
-     */
-    private $children;
-
-    /**
      * @var array
      */
     private $data = array();
@@ -51,10 +46,7 @@ class Scope
      */
     public function enter()
     {
-        $child = new self($this);
-        $this->children[] = $child;
-
-        return $child;
+        return new self($this);
     }
 
     /**