+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -4,8 +4,13 @@ CHANGELOG
 3.0.0
 -----
 
-* Removed `Symfony\Component\HttpKernel\Kernel::init()`
-* Removed `Symfony\Component\HttpKernel\Kernel::isClassInActiveBundle()` and `Symfony\Component\HttpKernel\KernelInterface::isClassInActiveBundle()`
+ * Removed `Symfony\Component\HttpKernel\Kernel::init()`
+ * Removed `Symfony\Component\HttpKernel\Kernel::isClassInActiveBundle()` and `Symfony\Component\HttpKernel\KernelInterface::isClassInActiveBundle()`
+
+2.7.0
+-----
+
+ * added the HTTP status code to profiles
 
 2.6.0
 -----